refactor: split cli command modules
